+++ src/components/NightStudy/NightStudyBan/NightStudyBanItem/Student/style.ts
@@ -1,0 +1,5 @@
+import {CSSObject} from "styled-components";
+
+export const StudentTD: CSSObject = {
+  width: "11.5%",
+}
