--- src/components/NightStudy/NightStudyBan/NightStudyBanItem/style.ts
@@ -1,13 +1,0 @@
-import {CSSObject} from "styled-components";
-
-export const StudentTBody: CSSObject = {
-  width: "100%",
-  fontWeight: "600px",
-  fontSize: "14px",
-  display: "flex",
-  flexDirection: "column"
-}
-
-export const StudentTD: CSSObject = {
-  width: "11.5%",
-}
