+++ src/components/NightStudy/NightStudyBan/index.tsx
@@ -2,7 +2,6 @@ import * as S from "./style";
 import NightStudyBanHeader from "./NightStudyBanHeader";
 import TableAttribute from "components/common/TableAttribute";
 import {NIGHTSTUDY_BAN_TABLE_ITEMS} from "constants/LateNight/latenight.constant";
-import ErrorBoundary from "components/common/ErrorBoundary";
 import React, {Suspense} from "react";
 import SkeletonComponent from "components/common/Skeleton";
 import NightStudyBanItem from "./NightStudyBanItem";
@@ -13,6 +12,7 @@ import {
   NightStudySelectGradeAtom
 } from "stores/NightStudy/nightstudy.store";
 import NightStudyBanModal from "./NightStudyBanModal";
+import {DodamErrorBoundary} from "@b1nd/dds-web";
 
 const NightStudyBan = () => {
   const searchValue = useRecoilValue(NightStudySearchAtom);
@@ -26,7 +26,7 @@ const NightStudyBan = () => {
         constant={NIGHTSTUDY_BAN_TABLE_ITEMS}
         thStyle={{width: "11%"}}
       >
-        <ErrorBoundary text="학생들을 불러오지 못했습니다." showButton={true}>
+        <DodamErrorBoundary text="학생들을 불러오지 못했습니다." showButton={true}>
           <Suspense fallback={<SkeletonComponent height={70} />}>
             <NightStudyBanItem
               searchValue={searchValue}
@@ -35,7 +35,7 @@ const NightStudyBan = () => {
             />
           </Suspense>
           <NightStudyBanModal/>
-        </ErrorBoundary>
+        </DodamErrorBoundary>
       </TableAttribute>
     </S.NightStudyBanContainer>
   )
